Resume learning rate scheduling.

+++ tests/integration/test_transformers.py
@@ -14,8 +14,9 @@ import shutil
 from parlai.scripts.train_model import TrainLoop, setup_args
 
 
-def _mock_train(**args):
-    outdir = tempfile.mkdtemp()
+def _mock_train(outdir=None, keepoutdir=False, **args):
+    if not outdir:
+        outdir = tempfile.mkdtemp()
     parser = setup_args()
     parser.set_defaults(
         model_file=os.path.join(outdir, "model"),
@@ -25,8 +26,8 @@ def _mock_train(**args):
     with contextlib.redirect_stdout(stdout):
         tl = TrainLoop(parser.parse_args(print_args=False))
         valid, test = tl.train()
-
-    shutil.rmtree(outdir)
+    if not keepoutdir:
+        shutil.rmtree(outdir)
     return stdout.getvalue(), valid, test
 
 
@@ -40,7 +41,7 @@ class TestTransformerRanker(unittest.TestCase):
             optimizer='adamax',
             learningrate=7e-3,
             batchsize=32,
-            num_epochs=10,
+            num_epochs=15,
             no_cuda=True,
             n_layers=1,
             n_heads=1,
@@ -49,13 +50,66 @@ class TestTransformerRanker(unittest.TestCase):
             scores_norm='dot',
         )
 
-        self.assertTrue(
-            valid['hits@1'] > 0.95,
+        self.assertGreaterEqual(
+            valid['hits@1'],
+            0.95,
             "valid hits@1 = {}\nLOG:\n{}".format(valid['hits@1'], stdout)
         )
-        self.assertTrue(
-            test['hits@1'] > 0.95,
+        self.assertGreaterEqual(
+            test['hits@1'],
+            0.95,
             "test hits@1 = {}\nLOG:\n{}".format(test['hits@1'], stdout)
+        )
+
+    def test_resuming(self):
+        outdir = tempfile.mkdtemp()
+        stdout1, valid1, test1 = _mock_train(
+            outdir=outdir,
+            keepoutdir=True,
+            task='integration_tests:CandidateTeacher',
+            model='transformer/ranker',
+            optimizer='adamax',
+            learningrate=1e-3,
+            batchsize=32,
+            num_epochs=1,
+            no_cuda=True,
+            n_layers=1,
+            n_heads=1,
+            ffn_size=32,
+            embedding_size=32,
+            scores_norm='dot',
+            warmup_updates=1,
+            lr_scheduler='invsqrt',
+        )
+        stdout2, valid2, test2 = _mock_train(
+            outdir=outdir,
+            keepoutdir=False,
+            task='integration_tests:CandidateTeacher',
+            model='transformer/ranker',
+            optimizer='adamax',
+            learningrate=1e-3,
+            batchsize=32,
+            num_epochs=1,
+            no_cuda=True,
+            n_layers=1,
+            n_heads=1,
+            ffn_size=32,
+            embedding_size=32,
+            scores_norm='dot',
+            warmup_updates=1,
+            lr_scheduler='invsqrt',
+        )
+        # make sure the number of updates is being tracked correctly
+        self.assertGreater(
+            valid2['num_updates'],
+            valid1['num_updates'],
+            'Number of updates is not increasing'
+        )
+        # make sure the learning rate is decreasing
+        self.assertLess(
+            valid2['lr'],
+            valid1['lr'],
+            'Learning rate is not decreasing'
         )
 
 
@@ -78,12 +132,14 @@ class TestTransformerGenerator(unittest.TestCase):
             rank_candidates=True,
         )
 
-        self.assertTrue(
-            valid['hits@1'] > 0.95,
+        self.assertGreaterEqual(
+            valid['hits@1'],
+            0.95,
             "valid hits@1 = {}\nLOG:\n{}".format(valid['hits@1'], stdout)
         )
-        self.assertTrue(
-            test['hits@1'] > 0.95,
+        self.assertGreaterEqual(
+            test['hits@1'],
+            0.95,
             "test hits@1 = {}\nLOG:\n{}".format(test['hits@1'], stdout)
         )
 
@@ -94,7 +150,7 @@ class TestTransformerGenerator(unittest.TestCase):
             optimizer='adamax',
             learningrate=7e-3,
             batchsize=32,
-            num_epochs=15,
+            num_epochs=20,
             no_cuda=True,
             n_layers=1,
             n_heads=1,
@@ -103,21 +159,75 @@ class TestTransformerGenerator(unittest.TestCase):
             beam_size=5,
         )
 
-        self.assertTrue(
-            valid['ppl'] < 1.20,
+        self.assertLessEqual(
+            valid['ppl'],
+            1.20,
             "valid ppl = {}\nLOG:\n{}".format(valid['ppl'], stdout)
         )
-        self.assertTrue(
-            valid['bleu'] >= .95,
+        self.assertGreaterEqual(
+            valid['bleu'],
+            0.95,
             "valid blue = {}\nLOG:\n{}".format(valid['bleu'], stdout)
         )
-        self.assertTrue(
-            test['ppl'] < 1.20,
+        self.assertLessEqual(
+            test['ppl'],
+            1.20,
             "test ppl = {}\nLOG:\n{}".format(test['ppl'], stdout)
         )
-        self.assertTrue(
-            test['bleu'] >= .95,
+        self.assertGreaterEqual(
+            test['bleu'],
+            0.95,
             "test bleu = {}\nLOG:\n{}".format(test['bleu'], stdout)
+        )
+
+    def test_resuming(self):
+        outdir = tempfile.mkdtemp()
+        stdout1, valid1, testr1 = _mock_train(
+            outdir=outdir,
+            keepoutdir=True,
+            task='integration_tests:NocandidateTeacher',
+            model='transformer/generator',
+            optimizer='adamax',
+            learningrate=1e-3,
+            batchsize=32,
+            num_epochs=1,
+            no_cuda=True,
+            n_layers=1,
+            n_heads=1,
+            ffn_size=32,
+            embedding_size=32,
+            skip_generation=True,
+            lr_scheduler='invsqrt',
+            warmup_updates=1,
+        )
+        stdout2, valid2, test2 = _mock_train(
+            outdir=outdir,
+            keepoutdir=False,
+            task='integration_tests:NocandidateTeacher',
+            model='transformer/generator',
+            optimizer='adamax',
+            learningrate=1e-3,
+            batchsize=32,
+            num_epochs=1,
+            no_cuda=True,
+            n_layers=1,
+            n_heads=1,
+            ffn_size=32,
+            embedding_size=32,
+            skip_generation=True,
+            lr_scheduler='invsqrt',
+        )
+        # make sure the number of updates is being tracked correctly
+        self.assertGreater(
+            valid2['num_updates'],
+            valid1['num_updates'],
+            'Number of updates is not increasing'
+        )
+        # make sure the learning rate is decreasing
+        self.assertLess(
+            valid2['lr'],
+            valid1['lr'],
+            'Learning rate is not decreasing'
         )
